+++ Soggins/docs/release-notes.md
@@ -4,6 +4,7 @@
 
 ## Future version
 
+* Added event dialogue.
 * For modders: Added an API to get Soggins' `ExternalDeck`.
 
 ## 1.1.3